+++ src/components/manage/Widgets/UrlWidget.jsx
@@ -73,15 +73,10 @@ export const UrlWidget = (props) => {
     newValue = isInternalURL(newValue) ? addAppURL(newValue) : newValue;
 
     if (!isInternalURL(newValue) && newValue.length > 0) {
-      if (URLUtils.isMail(URLUtils.normaliseMail(newValue))) {
-        newValue = URLUtils.normaliseMail(newValue);
-      } else if (URLUtils.isTelephone(newValue)) {
-        newValue = URLUtils.normalizeTelephone(newValue);
-      } else {
-        newValue = URLUtils.normalizeUrl(newValue);
-        if (!URLUtils.isUrl(newValue)) {
-          setIsInvalid(true);
-        }
+      const checkedURL = URLUtils.checkAndNormalizeUrl(newValue);
+      newValue = checkedURL.url;
+      if (!checkedURL.isValid) {
+        setIsInvalid(true);
       }
     }